+++ src/lib/common/sol-mainloop-impl-zephyr.c
@@ -1,7 +1,7 @@
 /*
  * This file is part of the Soletta Project
  *
- * Copyright (C) 2016 Intel Corporation. All rights reserved.
+ * Copyright (C) 2015 Intel Corporation. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,9 +16,12 @@
  * limitations under the License.
  */
 
+#include <stdbool.h>
+
 #include <nanokernel.h>
 
-#include "sol-log.h"
+#include "sol-mainloop-common.h"
+#include "sol-mainloop-impl.h"
 #include "sol-mainloop-zephyr.h"
 #include "sol-util.h"
 
@@ -34,12 +37,17 @@ static struct nano_fifo _sol_mainloop_free_events;
 #define EVENTS_BUF_SIZE (MAX_QUEUED_EVENTS * sizeof(struct me_fifo_entry))
 static struct me_fifo_entry _events[EVENTS_BUF_SIZE];
 
+static nano_thread_id_t main_thread_id;
+static struct nano_sem _sol_mainloop_lock;
+
 int
 sol_mainloop_impl_platform_init(void)
 {
     int i;
 
-    sol_mainloop_zephyr_common_init();
+    main_thread_id = sys_thread_self_get();
+    nano_sem_init(&_sol_mainloop_lock);
+    nano_sem_give(&_sol_mainloop_lock);
 
     nano_fifo_init(&_sol_mainloop_pending_events);
     nano_fifo_init(&_sol_mainloop_free_events);
@@ -51,6 +59,41 @@ sol_mainloop_impl_platform_init(void)
     }
 
     return 0;
+}
+
+void
+sol_mainloop_impl_lock(void)
+{
+    nano_sem_take(&_sol_mainloop_lock, TICKS_UNLIMITED);
+}
+
+void
+sol_mainloop_impl_unlock(void)
+{
+    nano_sem_give(&_sol_mainloop_lock);
+}
+
+bool
+sol_mainloop_impl_main_thread_check(void)
+{
+    return main_thread_id == sys_thread_self_get();
+}
+
+void
+sol_mainloop_impl_main_thread_notify(void)
+{
+    static const struct mainloop_event me = {
+        .cb = NULL,
+        .data = NULL
+    };
+
+    sol_mainloop_event_post(&me);
+}
+
+void
+sol_mainloop_impl_platform_shutdown(void)
+{
+    sol_mainloop_common_source_shutdown();
 }
 
 int
@@ -81,4 +124,39 @@ sol_mainloop_events_process(int32_t sleeptime)
             mfe->me.cb((void *)mfe->me.data);
         nano_task_fifo_put(&_sol_mainloop_free_events, mfe);
     } while ((mfe = nano_task_fifo_get(&_sol_mainloop_pending_events, TICKS_NONE)));
+}
+
+static inline int32_t
+ticks_until_next_timeout(void)
+{
+    struct timespec ts;
+    bool ret;
+
+    sol_mainloop_impl_lock();
+    ret = (sol_mainloop_common_idler_first() != NULL);
+    sol_mainloop_impl_unlock();
+    if (ret)
+        return TICKS_NONE;
+
+    sol_mainloop_impl_lock();
+    ret = sol_mainloop_common_timespec_first(&ts);
+    sol_mainloop_impl_unlock();
+    if (!ret)
+        return TICKS_UNLIMITED;
+
+    return ts.tv_sec * sys_clock_ticks_per_sec +
+        ((long long)sys_clock_ticks_per_sec * ts.tv_nsec) / NSEC_PER_SEC;
+}
+
+void
+sol_mainloop_impl_iter(void)
+{
+    int32_t sleeptime;
+
+    sol_mainloop_common_timeout_process();
+
+    sleeptime = ticks_until_next_timeout();
+    sol_mainloop_events_process(sleeptime);
+
+    sol_mainloop_common_idler_process();
 }
